requirements.txt
git+https://github.com/openai/CLIP.git
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
scikit_learn==1.3.2
streamlit==1.30.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
opencv==4.9.0

